moved everything to one header file.
